ref(references): extract the PHP reflection layer into PhpIndexBridge
PhpClassResolver mixed two concerns: turning a Phel interop symbol into a PHP FQN
and member name (Phel-side mapping), and reflectively querying the optional PHP
plugin for that FQN (the dangerous part, where a ProcessCanceledException arrives
wrapped in an InvocationTargetException and must be rethrown). Everything that
touches com.jetbrains.php.* now lives in PhpIndexBridge, so the resolver reads as
plain mapping.

  PhpClassResolver   356 -> 140   Phel symbol -> FQN / member name
  PhpIndexBridge            190   the reflective PhpIndex lookup, in one place

Also folds readStringProperty/readBoolProperty into one readProperty (they were
byte-identical but for the cast), and drops the manual loadPhpIndexClass /
phpIndexClass!=null dance at three call sites into a single isAvailable().
rethrowIfPlatformControlFlow moves with the reflection it guards; its test now
targets the bridge.

Behaviour preserved: 1143 tests, 0 failures.
